infra/vision/requirements.txt
flask==3.0.3
numpy==1.26.4
opencv-python-headless==4.10.0.84
ultralytics==8.2.103
Pillow==10.4.0
